### 2023-03-08 #5233 @danielskatz
👋 @VanillaBrooks - thanks for your submission.  While we look for an editor, you could work on the possibly missing DOIs that editorialbot suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.

I think you will also need to talk about other systems that have somewhat comparable functionality in your paper as well, such as many workflow systems.
