+++ pybind/README.md
@@ -5,13 +5,8 @@ Python bindings to `distribute` for automated creation of configuration files
 ## Documentation
 
 Python documentation is available 
-<<<<<<< HEAD
 [here](https://fluid-dynamics-group.github.io/distribute/python/).
-=======
-[here](https://fluid-dynamics-group.github.io/distribute-docs/python/).
->>>>>>> fdfb12ebdaa57f3a6bf85d067046e8781bec9c71
-Python installation instructions are available
-[here](file:///home/brooks/github/fluids/distribute/docs/book/install.html#python-api-install)
+Python installation instructions are available [here](https://fluid-dynamics-group.github.io/distribute/install.html#python-api-install).
 
 ## Developer Documentation
 
