### 2023-03-14 a58f679
add doi for slurm and reuther papers
