+++ paper/paper.bib
@@ -12,7 +12,8 @@
   booktitle={Job Scheduling Strategies for Parallel Processing: 9th International Workshop, JSSPP 2003, Seattle, WA, USA, June 24, 2003. Revised Paper 9},
   pages={44--60},
   year={2003},
-  organization={Springer}
+  organization={Springer},
+  url={doi.org/10.1007/10968987_3}
 }
 
 @inproceedings{reuther2016scheduler,
@@ -21,5 +22,6 @@
   booktitle={2016 IEEE High Performance Extreme Computing Conference (HPEC)},
   pages={1--6},
   year={2016},
-  organization={IEEE}
+  organization={IEEE},
+  url={doi.org/10.1109/hpec.2016.7761604}
 }
