### 2023-03-14 5d99881
add https to url, include specific doi field
