+++ paper/paper.bib
@@ -13,7 +13,8 @@
   pages={44--60},
   year={2003},
   organization={Springer},
-  url={doi.org/10.1007/10968987_3}
+  url={https://doi.org/10.1007/10968987_3},
+  doi={10.1007/10968987_3}
 }
 
 @inproceedings{reuther2016scheduler,
@@ -23,5 +24,6 @@
   pages={1--6},
   year={2016},
   organization={IEEE},
-  url={doi.org/10.1109/hpec.2016.7761604}
+  url={https://doi.org/10.1109/hpec.2016.7761604},
+  doi={10.1109/hpec.2016.7761604}
 }
