### 2023-03-14 #5233 @VanillaBrooks
I agree that the paper could be longer. Mostly I was trying to meet the requirements [here](https://joss.readthedocs.io/en/latest/review_checklist.html#software-paper) while avoiding technical / implementation details that would escape a general audience.

I think that there is a lot of interesting things to talk about with regards to automatic output archival and selective querying and download of results that make this software useful. Would these be out of scope for this paper?

We have been using this project extensively in our lab to reduce downtime on compute resources for computational fluid dynamics (CFD) simulations. However, none of these efforts have been published (academically or source code). Is this worth the mention?

As you mention, differences with other workload managers / schedulers can definitely be discussed more.
