### 2023-03-27 0d4fab2
rework apptainer documentation section to be more clear and robust
