+++ docs/src/SUMMARY.md
@@ -4,7 +4,12 @@
 - [Installation](./install.md)
 - [Overview](./overview.md)
 - [Configuration](./configuration.md)
-	- [Apptainer Jobs](./apptainer.md)
+	- [Apptainer Configuration](./apptainer.md)
+		- [Apptainer Overview](./apptainer-introduction.md)
+		- [Mutable Filesystem](./apptainer-mutable-filesystem.md)
+		- [Input Files](./apptainer-input-files.md)
+		- [Packaging Solvers](./apptainer-packaging-solvers.md)
+		- [Running Locally](./apptainer-running-locally.md)
 	- [Python Jobs](./python.md)
 - [Command Basics](./commands.md)
 - [Python Api](./python_api.md)
