### 2023-03-27 a3295b7
rework python documentation
