+++ docs/src/SUMMARY.md
@@ -8,9 +8,15 @@
 		- [Apptainer Overview](./apptainer-introduction.md)
 		- [Mutable Filesystem](./apptainer-mutable-filesystem.md)
 		- [Input Files](./apptainer-input-files.md)
+		- [Storing Results](./apptainer-storing-results.md)
 		- [Packaging Solvers](./apptainer-packaging-solvers.md)
 		- [Running Locally](./apptainer-running-locally.md)
 	- [Python Jobs](./python.md)
+		- [Build Scripts](./python-build-scripts.md)
+		- [Input Files](./python-input-files.md)
+		- [Storing Results](./python-storing-results.md)
+		- [Execution Scripts](./python-execution-scripts.md)
+		- [Full Example](./python-full-example.md)
 - [Command Basics](./commands.md)
 - [Python Api](./python_api.md)
 - [Available Capabilities](capabilities.md)
