+++ docs/src/python.md
@@ -20,7 +20,8 @@ meta:
     - apptainer
 python:
   initialize:
-    build_file: /path/to/build.py
+    build_file: 
+	  path: /path/to/build.py
     required_files:
       - path: /file/always/present/1.txt
         alias: optional_alias.txt
@@ -30,7 +31,8 @@ python:
         alias: ~
   jobs:
     - name: job_1
-      file: execute_job.py
+      file: 
+	    path: execute_job.py
       required_files:
         - path: job_configuration_file.json
           alias: ~
@@ -38,336 +40,114 @@ python:
           alias: input.json
 ```
 
-## What You Are Provided
+### `meta` section
 
-You may ask, what do your see when they are executed on a node? While the base folder structure remains the same,
-the files you are provided differ. Lets say you are executing the following section of a configuration file:
+The `meta` section is identical to the `meta` section of python. 
 
-```yaml
-python:
-  initialize:
-    build_file: /path/to/build.py
-    required_files:
-      - path: file1.txt
-      - path: file999.txt
-        alias: file2.txt
+The `namespace` field is a string that describes all of the `batch_name`s that 
+are a part of this `namespace`.
+
+The `batch_name` is an a simple string to identify what jobs are in this batch.
+
+If your `distribute` cluster has been configured by your adminstrator for matrix support, 
+you can specify the `matrix` field to be your matrix username in the form of `@your-user:matrix.org`
+and you will receive a message when all of the jobs in this job batch finish.
+
+The `capabilities` section is a list of dependencies that are required to be present on the compute node.
+Since you are responsible for bundling all your dependencies in your apptainer image, the only capabilities
+requires here are `apptainer`. If you require a GPU on the compute node, you can also specify the `gpu`
+capability.
+
+On the server, each `batch_name` must be unique to each `namespace`. This means the results of submitting
+two batches of batch name `name_xyz` with identical namespaces is undefined. This is because on
+the server each batch name is a subdirectory of the namespace. For example, if we have run
+two job batches `job_batch_01` and `job_batch_02` under the namespace `example_namespace`,
+the directory structure might look like 
+
+```
+example_namespace/
+├── job_batch_01
+│   ├── job_1
+│   │   └── stdout.txt
+│   ├── job_2
+│   │   └── stdout.txt
+│   └── job_3
+│       └── stdout.txt
+└── job_batch_02
+    ├── job_1
+    │   └── stdout.txt
+    ├── job_2
+    │   └── stdout.txt
+    └── job_3
+        └── stdout.txt
+```
+
+For example, if I am running many aeroacoustic studies with OpenFoam, a namespace might be 
+`austin_openfoam_aeroacoustic`, and then I run batches with incremented names,
+`supersonic_aeroacoustic_00X`, `subsonic_aeroacoustic_00X`, `transsonic_aeroacoustic_00X`. My `meta`
+section for one of these job batches would look like:
+
+```
+meta:
+  batch_name: subsonic_aeroacoustic_004
+  namespace: austin_openfoam_aeroacoustic
+  matrix: @my-username:matrix.org
+  capabilities:
+    - apptainer
+    - gpu
+```
+
+### `python` section
+
+the `python` section has two subsections: `initialize` which handles compiling and setting
+up all code for execution, and `jobs` which specifies a list of jobs that are part of this job batch.
+
+### `initialize` section
+
+The `build_file` key contains a `path` to a python file.
+This python file is responsible for compiling and setting up an environment on the system
+to execute all of the jobs in the job batch.
+
+The `required_files` contains a list of files, each file with a `path` and `alias`.
+The `path` is a path to a file on disk, and the `alias` is a filename as it will appear
+in the `/input` directory when compiling all related tasks. 
[…]
