### 2023-03-27 030ff4f
update capabilities docs
