+++ docs/src/python-full-example.md
@@ -1,1 +1,4 @@
 # Full Example
+
+A simpler example of a python job has been compiled and verified 
+[here](https://github.com/Fluid-Dynamics-Group/distribute/tree/cancel-tasks/examples/python).
