+++ Cargo.toml
@@ -16,7 +16,7 @@ derive_more = "0.99.16"
 bincode = { version = "1.3.3", optional = true }
 walkdir = { version = "2.3.2", optional = true }
 tokio = { version = "1.24.2", features = ["rt", "macros", "net", "io-util", "time", "sync", "fs", "process", "rt-multi-thread"], optional = true } 
-chrono = { version = "0.4.23", optional = true }
+chrono = { version = "0.4.24", optional = true }
 futures = { version = "0.3.25", optional = true}
 nix = { version = "0.26.1", optional = true }
 num_cpus = { version = "1.15.0", optional = true }
@@ -33,7 +33,7 @@ getset = "0.1.2"
 [dependencies.matrix-notify]
 default_features = false
 git = "https://github.com/fluid-Dynamics-Group/matrix-notify"
-rev = "f3d4def0b1386ab5ee686857a029fa5716151a00"
+rev = "d88dd6518ec3ce1e5bb1992ecf1e229be5b98016"
 
 [dependencies.pyo3]
 version = "0.17.3" 
