+++ src/server/schedule.rs
@@ -593,9 +593,12 @@ mod tests {
             None,
         );
 
-        let jgpu = config::python::Job::new("jgpu".into(), File::new_relative("some_path"), vec![], None);
-        let j1 = config::python::Job::new("j1".into(), File::new_relative("some_path"), vec![], None);
-        let j2 = config::python::Job::new("j2".into(), File::new_relative("some_path"), vec![], None);
+        let jgpu =
+            config::python::Job::new("jgpu".into(), File::new_relative("some_path"), vec![], None);
+        let j1 =
+            config::python::Job::new("j1".into(), File::new_relative("some_path"), vec![], None);
+        let j2 =
+            config::python::Job::new("j2".into(), File::new_relative("some_path"), vec![], None);
 
         let build = config::python::Initialize::new(
             config::common::File::new_relative("build_file"),
@@ -611,15 +614,17 @@ mod tests {
         let cpu_description = config::python::Description::new(build.clone(), vec![j1, j2]);
         let gpu_description = config::python::Description::new(build.clone(), vec![jgpu]);
 
-        let cpu_config = config::Jobs::from(config::PythonConfig::new(cpu_meta, cpu_description, None))
-            .hashed()
-            .unwrap()
-            .into();
+        let cpu_config =
+            config::Jobs::from(config::PythonConfig::new(cpu_meta, cpu_description, None))
+                .hashed()
+                .unwrap()
+                .into();
 
-        let gpu_config = config::Jobs::from(config::PythonConfig::new(gpu_meta, gpu_description, None))
-            .hashed()
-            .unwrap()
-            .into();
+        let gpu_config =
+            config::Jobs::from(config::PythonConfig::new(gpu_meta, gpu_description, None))
+                .hashed()
+                .unwrap()
+                .into();
 
         (cpu_config, gpu_config, scheduler)
     }
