+++ paper/paper.bib
@@ -27,3 +27,12 @@
   url={https://doi.org/10.1109/hpec.2016.7761604},
   doi={10.1109/hpec.2016.7761604}
 }
+
+@misc{apptainer,
+  author = {Apptainer Collaboration},
+  title = {Apptainer: Application containers for Linux },
+  year = {2022},
+  publisher = {GitHub},
+  journal = {GitHub repository},
+  url = {https://github.com/apptainer/apptainer}
+}
