+++ paper/paper.md
@@ -43,27 +43,75 @@ others remain unused. Dividing simulations manually among computers is time-cons
 ineffective due to varying hardware configurations. Additionally, multiple individuals running 
 jobs on a single computer can slow down progress for everyone involved.
 
-Existing workload managers such as SLURM [@SLURM] and TORQUE [@TORQUE]
-are useful for cluster computing but require specialized hardware not typically available on 
-standard workstations. Furthermore, these workload managers assume that nodes are solely 
-dedicated to computing, which may not always be the case.
+To meet the growing challenges of managing compute resources between 
+researchers, workload managers such as Slurm [@SLURM] and TORQUE [@TORQUE] quickly 
+rose in popularity. These workload managers enforce strict upper bounds on memory,
+CPU time, and CPU cores on each job submitted. Leveraging known runtime and hardware constraints, 
+these traditional workload managers optimize the scheduling of hundreds of jobs concurrently
+across thousands of CPU cores and terrabytes of memory.
+
+While Slurm and TORQUE provide elegant and powerful solutions to scientific computing,
+they are incompatible with cheap off-the-shelf hardware present in many research labs.
+In order to schedule a compute task across hundreds of cores and dozens of
+individual computers concurrently, traditional workload managers require specialized 
+memory and filesystems to enable communication between multiple computers. 
+Although this hardware is standard in conventional supercomputer clusters, 
+it leaves collections of common desktop computers present in research labs underserved.
+
+Users with access to scientific computing clusters are usually given minimal priviledges:
+only enough to start jobs and download the results. Since users do not have 
+access to individual compute nodes, Slurm and TORQUE generally assume 
+they are used for purely compute reasons and provide no method to temporarily halt
+a job on a single node. However, in a lab environment, desktops used for compute purposes 
+often double as workstations for researchers, and the execution of jobs in the background
+may slow down day-to-day work.
+
+# *distribute* Overview
 
 Our tool, *distribute*, eliminates the need for specialized hardware such as memory fabric 
 or shared filesystems, making almost no assumptions about the architecture of the computers. 
 It schedules multiple jobs concurrently across several computers, as shown in
 \autoref{fig:computing-layout}, avoiding the simultaneous execution of a single job across 
-multiple computers. While using *distribute*, normal workstation usage may slow down 
-during simulations, but we offer a way to temporarily pause background jobs to allow for normal work to continue.
-
-Running a multidimensional parameter space sweep may result in hundreds of jobs, and our tool provides a 
-Python package to programmatically generate the configuration file. Furthermore, *distribute* can 
-transpile its job execution configuration to the SLURM format, making it easy to run a batch of 
-jobs previously executed on a *distribute* cluster of in-house machines on a larger cluster with hundreds of cores.
+multiple computers. 
+Since the hardware that *distribute* targets also doubles as research workstations, 
+we offer a simple interface to pausing the background execution of jobs for normal 
+research work to continue.
 
 ![
 Left: traditional horizontally scaling compute framework. Right: proposed compute framework.
 \label{fig:computing-layout}
 ](./node_layout.png)
+
+While Slurm and TORQUE provide simple interfaces to schedule a single large job, *distribute* focuses 
+on the scheduling of tens to hundreds of small to medium sized jobs. In a *distribute* configuration
+file, a method for compiling a computational task ("solver") is specified with a list of job
+names. Each job name specifies a list of files that serve as inputs to the solver. 
+\autoref{fig:stateless-solver-formula} provides an overview of the *distribute* execution model.
+When each job is scheduled, *distribute* ensures that the solver is correctly compiled on 
+each compute node and provided with access to the input files for the job. When the job
+is completed, the solver outputs are transported and archived on the head node.
+
[…]
