### 2023-03-29 ec97634
remove paper/media from tracked files
