### 2023-03-31 08f52aa
fix capabilities doc typo
