### 2023-04-06 22a196b
bump cargo.toml version to 0.14.2, ensure hashed files use meta section
