+++ src/lib.rs
@@ -1,5 +1,4 @@
 #![doc = include_str!("../README.md")]
-
 #![allow(dead_code)]
 #![deny(missing_docs)]
 
