### 2023-04-10 4e1e251
test to ensure correct hashing w/ meta, better logging of node_meta for easier searching docs, better log file output without ansi garbage, log files required for a job
