+++ src/protocol/executing.rs
@@ -192,6 +192,10 @@ impl Machine<Executing, ClientExecutingState> {
             job_name,
             folder_state,
             cancel_addr,
+            node_meta: server::pool_data::NodeMetadata {
+                node_name: "SERVER".into(),
+                node_address: *conn.addr(),
+            },
         };
 
         ClientSendFilesState { conn, extra }
@@ -206,11 +210,17 @@ impl Machine<Executing, ClientExecutingState> {
         } = self.state;
         debug!("moving client executing -> prepare build");
 
-        #[allow(unused_mut)]
         let mut conn = conn.update_state();
 
-        #[cfg(test)]
-        assert!(conn.bytes_left().await == 0);
+        if conn.bytes_left().await != 0 {
+            error!(
+                "connection was not empty - this is guaranteed to cause error in following steps!"
+            );
+            panic!(
+                "connection was not empty - this is guaranteed to cause error in following steps!"
+            );
+        }
+
         super::prepare_build::ClientPrepareBuildState {
             conn,
             working_dir,
@@ -304,6 +314,7 @@ impl Machine<Executing, ServerExecutingState> {
                 let finish_msg = server::pool_data::FinishJob {
                     ident: self.state.job_identifier(),
                     job_name: self.state.run_info.task.name().to_string(),
+                    node_meta: self.state.common.node_meta.clone(),
                 };
 
                 let mark_finished_msg = server::JobRequest::FinishJob(finish_msg);
@@ -398,11 +409,16 @@ impl Machine<Executing, ServerExecutingState> {
         );
         let ServerExecutingState { conn, common, .. } = self.state;
 
-        #[allow(unused_mut)]
         let mut conn = conn.update_state();
 
-        #[cfg(test)]
-        assert!(conn.bytes_left().await == 0);
+        if conn.bytes_left().await != 0 {
+            error!(
+                "connection was not empty - this is guaranteed to cause error in following steps!"
+            );
+            panic!(
+                "connection was not empty - this is guaranteed to cause error in following steps!"
+            );
+        }
 
         super::prepare_build::ServerPrepareBuildState { conn, common }
     }
