### 2023-04-10 0b68a2a
place test-only checks on all state changes, add test to ensure completion of jobs with same .sif file and different batch_name-s in metadata
