### 2023-04-11 0cc3290
patch up doc tests
