### 2023-04-17 eceef2e
fix cancellation test using active default port
