+++ tests/failed_keepalive/distribute-nodes.yaml
@@ -2,5 +2,6 @@ nodes:
   - ip: "0.0.0.0"
     transport_port: 9967
     keepalive_port: 9968
+    cancel_port: 8956
     capabilities: []
     name: "testname"

[…]
