### 2023-06-15 1fe9b43
bump dependency versions, no ansi escape codes in file outputs, attempt to use system time for formatting
