+++ Cargo.toml
@@ -9,25 +9,25 @@ crate-type = ["cdylib", "rlib"]
 
 [dependencies]
 thiserror = "1.0.38"
-serde = { version = "1.0.151", features = ["derive"] }
-serde_yaml = "0.9.16"
+serde = { version = "1.0.164", features = ["derive"] }
+serde_yaml = "0.9.21"
 serde_json = "^1"
 derive_more = "0.99.16"
 bincode = { version = "1.3.3", optional = true }
 walkdir = { version = "2.3.2", optional = true }
-tokio = { version = "1.24.2", features = ["rt", "macros", "net", "io-util", "time", "sync", "fs", "process", "rt-multi-thread"], optional = true } 
-chrono = { version = "0.4.24", optional = true }
-futures = { version = "0.3.25", optional = true}
+tokio = { version = "1.28.2", features = ["rt", "macros", "net", "io-util", "time", "sync", "fs", "process", "rt-multi-thread"], optional = true } 
+chrono = { version = "0.4.26", optional = true }
+futures = { version = "0.3.28", optional = true}
 nix = { version = "0.26.1", optional = true }
 num_cpus = { version = "1.15.0", optional = true }
-tokio-stream = { version = "0.1.11", features = ["fs"], optional = true}
+tokio-stream = { version = "0.1.14", features = ["fs"], optional = true}
 sha1 = { version = "0.10.5", optional = true }
 signal-hook = { version = "0.3.9", default_features = false, optional = true }
-regex = { version = "1.7.0", optional = true }
-clap = { version = "4.0.32", features = ["derive"], optional = true }
+regex = { version = "1.8.4", optional = true }
+clap = { version = "4.3.4", features = ["derive"], optional = true }
 base16 = "0.2.1"
 tracing = { version = "0.1.37", optional = true }
-tracing-subscriber = { version = "0.3.16", optional = true }
+tracing-subscriber = { version = "0.3.17", features = ["time", "std", "local-time"], optional = true }
 getset = "0.1.2"
 
 [dependencies.matrix-notify]
