+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "distribute"
-version = "0.14.3"
+version = "0.14.4"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
