### 2023-06-15 98ef3b7
use mimalloc allocator, bump crate version
