### 2023-06-22 cc8fa78
update install docs for head node / compute node
