+++ docs/src/compute_node_setup.md
@@ -6,32 +6,54 @@ First, ensure you have completed the common setup from the [admin setup](./admin
 
 create the following directories:
 
-```
+```bash
 mkdir $HOME/data
 mkdir $HOME/logs
 ```
 
-Ensure that `distrbute` is installed for this user (running `distribute` should 
-result in some output with no errors)
+Clone `distribute` from the github repo and install the identical version as the head node.
+
+```bash
+git clone https://github.com/Fluid-Dynamics-Group/distribute --depth 1
+cd distribute
+```
+
+then, to update to `$VERSION` (fish shell syntax):
+
+```bash
+# for fish shell
+set VERSION "0.14.5"
+
+git fetch -a
+git checkout release-$VERSION
+git pull
+
+cargo install --path .
+
+rm ~/logs/output.log
+systemctl restart distribute-compute
+```
+
+the most recent `$VERSION` is usually up to date [here](https://github.com/Fluid-Dynamics-Group/distribute/blob/master/install/update.sh)
 
 ## While Root
 
 Clone the repo with the correct version of `distribute` you are using
 
-```
+```bash
 git clone https://github.com/Fluid-Dynamics-Group/distribute --depth 1
 cd distribute
 ```
 
 copy the compute service to the system directory:
 
-```
+```bash
 sudo cp install/distribute-compute.service /etc/systemd/system/
 ```
 
 start the service and enable it at startup:
 
-```
+```bash
 sudo systemctl daemon-reload
 sudo systemctl enable distribute-compute
 sudo systemctl start distribute-compute
@@ -42,18 +64,25 @@ have to be updated with those paths since it relies on hard-coded paths.
 
 ## Updating
 
-To update, simply reinstall `distribute` and restart the systemd service. On the head node:
+To update, simply reinstall `distribute` and restart the systemd service. On the compute node (for a fixed version at the time of writing):
 
+```bash
+# for fish shell
+set VERSION "0.14.5"
+
+git fetch -a
+git checkout release-$VERSION
+git pull
+
+cargo install --path .
+
+rm ~/logs/output.log
+systemctl restart distribute-compute
 ```
+
+the most recent `$VERSION` is usually up to date [here](https://github.com/Fluid-Dynamics-Group/distribute/blob/master/install/update.sh).
+Then, restart the systemd service so that is uses the new version of distribute:
+
+```bash
 systemctl restart distribute-server
 ```
-
-`distribute` requires that all compute nodes and head node have the same version of `distribute` installed.
-You can verify that all nodes are correctly setup or updated correctly with 
-
-```
-sudo su distribute
-cd ~/server
-distribute node-status
-```
-
