+++ docs/src/head_node_setup.md
@@ -9,7 +9,7 @@ send jobs to the cluster. Ensure this is documented somewhere.
 
 when logged into the user `distribute`, run
 
-```
+```bash
 mkdir $HOME/server
 mkdir $HOME/server/results
 mkdir $HOME/server/tmp
@@ -28,20 +28,20 @@ result in some output with no errors)
 
 Clone the repo with the correct version of `distribute` you are using
 
-```
+```bash
 git clone https://github.com/Fluid-Dynamics-Group/distribute --depth 1
 cd distribute
 ```
 
 copy the server service to the system directory:
 
-```
+```bash
 sudo cp install/distribute-server.service /etc/systemd/system/
 ```
 
 start the service and enable it at startup:
 
-```
+```bash
 sudo systemctl daemon-reload
 sudo systemctl enable distribute-server
 sudo systemctl start distribute-server
@@ -52,8 +52,28 @@ have to be updated with those paths since it relies on hard-coded paths.
 
 ## Updating
 
-To update, simply reinstall `distribute` and restart the systemd service. On the compute node:
+To update, simply reinstall `distribute` and restart the systemd service on the head node. For a fixed version at the time
+of writing, this script with work:
 
-```
+```bash
+cd ~/distribute
+
+# for fish shell
+set VERSION "0.14.5"
+
+git fetch -a
+git checkout release-$VERSION
+git pull
+
+cargo install --path .
+
+rm ~/logs/output.log
 systemctl restart distribute-compute
+```
+
+the most recent `$VERSION` is usually up to date [here](https://github.com/Fluid-Dynamics-Group/distribute/blob/master/install/update.sh).
+
+
+```bash
+systemctl restart distribute-server
 ```
