+++ static/example-nodes.yaml
@@ -2,14 +2,17 @@
 # number compatability and running status
 nodes:
   - ip: "134.197.94.36"
-    capabilities: [gpu, cargo, singularity, apptainer, lab1, lab1-3, lab1-2, gpu]
+    capabilities: [cargo, singularity, apptainer, lab1, lab1-3, lab1-2, gpu]
     name: "lab1"
-  - ip: "134.197.27.105"
+  - ip: "134.197.94.249"
     capabilities: [cargo, singularity, apptainer, lab2, lab2-3, python3, lab1-2]
     name: "lab2"
   - ip: "134.197.95.113"
     capabilities: [cargo, fftw, python3, openfoam, singularity, apptainer, lab1-3, lab2-3, lab3, lab3-5, gpu]
     name: "lab3"
-  - ip: "134.197.27.69"
-    capabilities: [cargo, singularity, apptainer, lab5, lab3-5, gpu]
-    name: "lab5"
+#- ip: "134.197.27.69"
+#  capabilities: [cargo, singularity, apptainer, lab5, lab3-5, gpu]
+#  name: "lab5"
+  - ip: "134.197.95.153"
+    capabilities: [singularity, apptainer, gpu, lab6]
+    name: "lab6"
