### 2023-06-29 2a7e2fd
more specific compute node install docs
