+++ static/example-nodes.yaml
@@ -2,7 +2,7 @@
 # number compatability and running status
 nodes:
   - ip: "134.197.94.36"
-    capabilities: [cargo, singularity, apptainer, lab1, lab1-3, lab1-2, gpu]
+    capabilities: [cargo, singularity, apptainer, lab1, lab1-3, lab1-2]
     name: "lab1"
   - ip: "134.197.94.249"
     capabilities: [cargo, singularity, apptainer, lab2, lab2-3, python3, lab1-2]
@@ -16,3 +16,6 @@ nodes:
   - ip: "134.197.95.153"
     capabilities: [singularity, apptainer, gpu, lab6]
     name: "lab6"
+#- ip: "134.197.95.153"
+#  capabilities: [singularity, apptainer, gpu, lab7]
+#  name: "lab7"
