+++ src/config/podman.rs
@@ -1,0 +1,283 @@
+use super::NormalizePaths;
+use derive_more::Constructor;
+
+use serde::{Deserialize, Serialize};
+use std::path::PathBuf;
+
+use super::common;
+
+#[cfg(feature = "cli")]
+use super::hashing;
+
+#[cfg(feature = "cli")]
+use crate::client::execute::FileMetadata;
+
+use getset::Getters;
+
+#[derive(Debug, Clone, Deserialize, Serialize, Constructor, Getters)]
+#[serde(deny_unknown_fields)]
+/// initialization and job specification for podman job batches
+pub struct Description<FILE> {
+    #[getset(get = "pub(crate)", get_mut = "pub")]
+    /// initialization information on podman job
+    pub initialize: Initialize<FILE>,
+    #[getset(get = "pub(crate)")]
+    /// list of jobs to execute in the job batch
+    pub jobs: Vec<Job<FILE>>,
+}
+
+#[cfg(feature = "cli")]
+impl Description<common::File> {
+    pub(super) fn verify_config(&self) -> Result<(), super::ConfigVerificationError> {
+        // ensure the execution of the
+        podman_image_exists(&self.initialize.image)?;
+
+        self.initialize
+            .required_files
+            .iter()
+            .try_for_each(|f| f.exists_or_err())?;
+
+        self.jobs.iter().try_for_each(|job| {
+            job.required_files
+                .iter()
+                .try_for_each(|f| f.exists_or_err())
+        })?;
+
+        Ok(())
+    }
+
+    pub(crate) fn len_jobs(&self) -> usize {
+        self.jobs.len()
+    }
+
+    pub(super) fn hashed(
+        &self,
+        meta: &super::Meta,
+    ) -> Result<Description<common::HashedFile>, super::MissingFilename> {
+        let initialize = self.initialize.hashed(meta)?;
+
+        let jobs = self
+            .jobs
+            .iter()
+            .enumerate()
+            .map(|(job_idx, job)| job.hashed(job_idx, meta))
+            .collect::<Result<Vec<_>, super::MissingFilename>>()?;
+
+        let desc = Description { initialize, jobs };
+
+        Ok(desc)
+    }
+}
+
+#[cfg(feature = "cli")]
+impl Description<common::HashedFile> {
+    pub(super) fn sendable_files(&self, is_user: bool) -> Vec<FileMetadata> {
+        let mut files = Vec::new();
+
+        self.initialize.sendable_files(is_user, &mut files);
+
+        for job in self.jobs.iter() {
+            job.sendable_files(is_user, &mut files);
+        }
+
+        files
+    }
+}
+
+impl<FILE> NormalizePaths for Description<FILE>
+where
+    FILE: NormalizePaths,
+{
+    fn normalize_paths(&mut self, base: PathBuf) {
+        for file in self.initialize.required_files.iter_mut() {
+            file.normalize_paths(base.clone());
+        }
+
+        // for jobs
+        for job in self.jobs.iter_mut() {
+            for file in job.required_files.iter_mut() {
+                file.normalize_paths(base.clone())
+            }
+        }
+    }
+}
+
+#[derive(Debug, Clone, Deserialize, Serialize, Constructor, getset::Getters)]
+#[serde(deny_unknown_fields)]
+/// initialization information for python jobs
+pub struct Initialize<FILE> {
+    #[getset(get = "pub(crate)")]
+    /// path to the image on a registry.
+    ///
+    /// Example: `docker.io/jellyfin/jellyfin:latest`
+    pub image: String,
+    #[serde(default = "Default::default")]
+    #[getset(get = "pub(crate)")]
+    /// list of required files used in podman initialization. Generally, for podman jobs,
+    /// these are simply files that will always be present
+    pub required_files: Vec<FILE>,
[…]
