+++ src/client/mod.rs
@@ -6,6 +6,7 @@ pub(crate) use utils::WorkingDir;
 pub(crate) use execute::{
     initialize_apptainer_job, initialize_python_job, run_apptainer_job, run_python_job,
     BindingFolderState,
+    initialize_podman_job
 };
 
 use crate::error::Error;
