+++ src/server/pool_data.rs
@@ -68,6 +68,9 @@ fn enumerate_paths(node_meta: &NodeMetadata, task_info: &TaskInfo) {
                 config::Init::Apptainer(app) => {
                     app.sendable_files(false, &mut files);
                 }
+                config::Init::Podman(podman) => {
+                    podman.sendable_files(false, &mut files);
+                }
             }
         }
     }
