+++ src/client/mod.rs
@@ -4,7 +4,7 @@ pub(crate) mod utils;
 pub(crate) use utils::WorkingDir;
 
 pub(crate) use execute::{
-    initialize_apptainer_job, initialize_python_job, run_apptainer_job, run_python_job,
+    initialize_apptainer_job, initialize_python_job, run_apptainer_job, run_python_job, run_podman_job,
     BindingFolderState,
     initialize_podman_job
 };
