### 2023-07-03 edb29a2
move to docker instead of podman - easier gpu acceleration
