### 2023-07-07 fd44458
state machine now tracks build information
