+++ src/protocol/send_files/mod.rs
@@ -135,12 +135,26 @@ async fn transport_files_with_large_file() {
     let server_conn = transport::Connection::from_connection(server_conn);
     let client_conn = transport::Connection::from_connection(client_conn);
 
+    let init = config::Init::Docker(config::docker::Initialize {
+        image: "docker.io/ubuntu:22.04".into(),
+        required_files: vec![],
+        required_mounts: vec![]
+    });
+
+    let build_info = server::pool_data::BuildTaskInfo {
+        namespace: "some namespace".into(),
+        batch_name: "some batch".into(),
+        identifier: server::JobSetIdentifier::Identity(0),
+        init
+    };
+
     let extra = SenderFinalStore {
         working_dir: base_dir.clone(),
         job_name: "test job name".into(),
         folder_state: client::BindingFolderState::new(),
         cancel_addr,
         node_meta: NodeMetadata::by_name("SERVER"),
+        build_info
     };
 
     let client_state = SenderState {
