### 2023-07-08 ffc9a84
docker setup documentation
