+++ docs/src/compute_node_setup.md
@@ -36,6 +36,16 @@ systemctl restart distribute-compute
 
 the most recent `$VERSION` is usually up to date [here](https://github.com/Fluid-Dynamics-Group/distribute/blob/master/install/update.sh)
 
+### Verifying installation
+
+Enter the `distribute` source code and run library tests with `cargo`. Note that you
+will have to complete the `docker` setup described below for all tests to pass.
+
+```bash
+cd ~/distribute
+cargo test --lib
+```
+
 ## While Root
 
 Clone the repo with the correct version of `distribute` you are using
@@ -61,6 +71,29 @@ sudo systemctl start distribute-compute
 
 Note that if you have deviated from username or folder structure above, `distribute-compute.service` will
 have to be updated with those paths since it relies on hard-coded paths.
+
+### Docker Setup
+
+As root, you must allow `docker` to be executed by the `distribute` user without `sudo`. This is
+done because the `distribute` executable explicitly avoids running with root privileges. Without 
+allowing `docker` to run rootless, you would be unable to start `docker` containers without distribute
+being run as root. See [this documentation](https://askubuntu.com/questions/477551/how-can-i-use-docker-without-sudo#answer-477554)
+on the security implications of this.
+
+As root, create `docker` group:
+
+```bash
+sudo groupadd docker
+```
+
+add the user `distribute` to the docker group
+
+```bash
+sudo gpasswd -a distribute docker
+newgrp docker
+```
+
+and restart the machine.
 
 ## Updating
 
