### 2023-07-10 19ed407
transport / configuration variants for docker
