+++ tests/docker_square/distribute-nodes.yaml
@@ -1,0 +1,1 @@
+../check_deallocate_jobs/distribute-nodes.yaml
