+++ tests/check_deallocate_jobs/distribute-nodes.yaml
@@ -2,5 +2,5 @@ nodes:
   - ip: "0.0.0.0"
     transport_port: 9967
     keepalive_port: 9968
-    capabilities: []
+    capabilities: ["docker", "apptainer"]
     name: "testname"
