### 2023-07-10 4d07e57
docker config now uses Image type to ensure stability
