### 2023-09-12 #5843 @editorialbot
[REVIEW]: distribute: Easy to use Resource Manager for Distributed Computing without Assumptions
**Submitting author:** @VanillaBrooks (<a href="http://orcid.org/0000-0002-1431-5199">Brooks</a>)
**Repository:** https://github.com/fluid-Dynamics-Group/distribute
**Branch with paper.md** (empty if default branch):
**Version:** 0.14.0
