### 2023-09-25 Fluid-Dynamics-Group/distribute#41 @HaoZeke
JOSS Review: Contributions 
The repo has no contributions from the second author listed? Part of https://github.com/openjournals/joss-reviews/issues/5843.

![image](https://github.com/Fluid-Dynamics-Group/distribute/assets/4336207/dcbb982b-00b9-4fd2-9ff3-0ef98dfa4b7e)
![image](https://github.com/Fluid-Dynamics-Group/distribute/assets/4336207/04d93c1a-2955-4706-be6d-a4ebb4f56610)
