### 2023-10-18 Fluid-Dynamics-Group/distribute#41 @VanillaBrooks
Hey @HaoZeke thanks for flagging this, its good attention to detail!

I would first like to apologize for my delay - I got caught up with deadlines at work & then have been sick over the past week. Thank you for your patience.

With regards to authorship, I was the sole author of the software in the sense of who-wrote-code on the project, but Aditya Nair provided invaluable guidance on the research problems this project needs to solve, as well as iterative feedback on usability and robustness of the software. I would be remiss if he did was not credited as a contributor, although he & I both agree he should be marked as a non-equal contributor.

I think this follows the joss guidelines on this [here](https://joss.readthedocs.io/en/latest/submitting.html?highlight=equal-contrib#authorship)

> Purely financial (such as being named on an award) and organizational (such as general supervision of a research group) contributions are not considered sufficient for co-authorship of JOSS submissions, but active project direction and other forms of non-code contributions are. The authors themselves assume responsibility for deciding who should be credited with co-authorship, and co-authors must always agree to be listed. In addition, co-authors agree to be accountable for all aspects of the work, and to notify JOSS if any retraction or correction of mistakes are needed after publication.

One thing I am confused on: it seems that the paper pipeline is somehow ignoring the existence of the `equal-contrib` key for each author. This marks myself and Aditya as equal authors in the rendered pdf. The relevant section in [the `paper.md` file](https://raw.githubusercontent.com/Fluid-Dynamics-Group/distribute/master/paper/paper.md) (github seems to not allow linking line numbers to markdown files):

```
authors:
  - name: Brooks Karlik
    orcid: 0000-0002-1431-5199
    equal-contrib: true
    corresponding: true
    affiliation: 1
  - name: Aditya Nair
    equal-contrib: false
    orcid: 0000-0002-8979-8420
    affiliation: 1
```

I am following the joss template [here](https://joss.readthedocs.io/en/latest/submitting.html?highlight=equal-contrib#example-paper-and-bibliography) for my knowledge on `equal-contrib`. Any ideas how to resolve this?
