+++ tests/apptainer_local/build.sh
@@ -1,3 +1,13 @@
+#! /usr/bin/bash
+
+# this script should be executed in its directory!
+# warning: this will create a folder at ~/apptainer/ and use it as a temporary directory
+# for building the container. The reasoning for this is detailed in the apptainer
+# documentation, but it it boils down to `/tmp` on linux being backed by an in-memory filesystem
+# that apptainer will often exceed: https://apptainer.org/docs/user/main/build_env.html#temporary-folders
+#
+# this folder is safe to remove once the build operation is complete
+
 OUT_SIF="apptainer_local.sif"
 rm $OUT_SIF
 
