### 2023-10-19 69e4427
break logging functionality to another module
