+++ src/lib.rs
@@ -39,6 +39,8 @@ mod slurm;
 mod template;
 #[cfg(feature = "cli")]
 mod transport;
+#[cfg(feature = "cli")]
+mod logging;
 
 #[cfg(feature = "cli")]
 use prelude::*;
@@ -49,9 +51,6 @@ pub use error::{CreateFile, Error, LogError, RunErrorLocal};
 #[macro_use]
 #[cfg(feature = "cli")]
 extern crate tracing;
-
-#[cfg(feature = "cli")]
-use tracing::level_filters::LevelFilter;
 
 pub use config::*;
 
@@ -90,143 +89,12 @@ mod reexports {
 #[cfg(feature = "cli")]
 /// create a logger instance sending output only to stdout
 pub fn logger() {
-    logger_cfg(LoggingOutput::Stdout, true);
+    logging::logger_cfg(logging::LoggingOutput::Stdout, true, true);
 }
 
-#[cfg(feature = "cli")]
-/// set locations for log outputs (stdout, file, both, or none)
-pub enum LoggingOutput {
-    /// log to stdout only
-    Stdout,
-    /// log to stdout *and* a file
-    StdoutAndFile(fs::File),
-    /// log exclusively a file
-    File(fs::File),
-    /// do not log
-    None,
-}
-
-#[cfg(feature = "cli")]
-impl LoggingOutput {
-    /// generate a [`LevelFilter`] from [`LoggingOutput`]
-    fn level(&self) -> LevelFilter {
-        match self {
-            Self::None => LevelFilter::OFF,
-            _ => LevelFilter::DEBUG,
-        }
-    }
-}
-
-#[cfg(feature = "cli")]
-/// setup logging with a specific [`LoggingOutput`] configuration
-///
-/// ## Parameters
-///
-/// `with_filename`: enable filename in logs
-pub fn logger_cfg(logging_output: LoggingOutput, with_filename: bool) {
-    use tracing_subscriber::fmt::time;
-
-    let time = time::SystemTime;
-
-    let stdout = std::io::stdout;
-    let logging_level = tracing::Level::DEBUG;
-
-    match logging_output {
-        LoggingOutput::None => (),
-        LoggingOutput::Stdout => {
-            let subscriber = tracing_subscriber::FmtSubscriber::builder()
-                .with_max_level(logging_level)
-                .with_file(with_filename)
-                .with_writer(stdout)
-                .with_ansi(true)
-                .with_timer(time)
-                .finish();
-
-            tracing::subscriber::set_global_default(subscriber)
-                .expect("setting default subscriber failed");
-        }
-        LoggingOutput::StdoutAndFile(file_writer) => {
-            use tracing_subscriber::{fmt, prelude::*, registry::Registry};
-
-            let stdout_subscriber = fmt::Layer::new()
-                .with_file(with_filename)
-                .with_writer(stdout.with_max_level(logging_level))
-                .with_timer(time.clone())
-                .with_ansi(false);
-
-            let file_subscriber = fmt::Layer::new()
-                .with_file(with_filename)
-                .with_writer(file_writer.with_max_level(logging_level))
-                .with_timer(time)
-                .with_ansi(false);
-
-            let subscriber = Registry::default()
-                .with(file_subscriber)
-                .with(stdout_subscriber);
-
-            tracing::subscriber::set_global_default(subscriber)
-                .expect("setting default subscriber failed");
-        }
-        LoggingOutput::File(file_writer) => {
-            let subscriber = tracing_subscriber::FmtSubscriber::builder()
-                .with_max_level(logging_level)
-                .with_file(with_filename)
[…]
