### 2023-10-19 72856f8
cargo fmt + allow running logging unit tests simulaneously
