+++ tests/apptainer_local.rs
@@ -1,6 +1,6 @@
-//! tests specific to apptainer execution. 
+//! tests specific to apptainer execution.
 //!
-//! These tests require that the correct `apptainer.sif` file was generated at 
+//! These tests require that the correct `apptainer.sif` file was generated at
 //! `./tests/apptainer_local/apptainer.sif`. The generation of this file is handled
 //! by running the shell script at .tests/apptainer_local
 
@@ -12,7 +12,7 @@ use std::fs;
 use std::path::PathBuf;
 
 #[tokio::test(flavor = "multi_thread", worker_threads = 2)]
-#[cfg(feature="test-apptainer")]
+#[cfg(feature = "test-apptainer")]
 async fn verify_apptainer_execution() {
     let dir = PathBuf::from("./tests/apptainer_local");
 
@@ -41,7 +41,7 @@ async fn verify_apptainer_execution() {
 }
 
 #[test]
-#[cfg(feature="test-apptainer")]
+#[cfg(feature = "test-apptainer")]
 fn slurm_output_verify() {
     if false {
         distribute::logger();
