+++ src/logging.rs
@@ -3,8 +3,16 @@
 //! access to this module is usually done through [`crate::logger`]. The `cli` feature must be
 //! enabled for this module to be compiled
 
-use tracing::level_filters::LevelFilter;
 use crate::prelude::*;
+use tracing::level_filters::LevelFilter;
+
+/// create a logger instance sending output only to stdout
+///
+/// helper function to setup logging for some unit and integration tests. More
+/// fine grained control of logging can be done with [`logger_cfg`]
+pub fn default() {
+    logger_cfg::<ThreadLocal>(LoggingOutput::Stdout, true);
+}
 
 /// set locations for log outputs (stdout, file, both, or none)
 pub enum LoggingOutput {
@@ -28,12 +36,54 @@ impl LoggingOutput {
     }
 }
 
+/// generic constructor-trait for setting up either thread-local logging or global logging
+///
+/// To setup global logging, see [`Global`], and for thread-local logging see [`ThreadLocal`]
+pub trait LogThreading {
+    /// construct the logger from the provided [`tracing`] subscriber
+    fn from_subscriber<S>(subscriber: S)
+    where
+        S: tracing::Subscriber + Send + Sync + 'static;
+}
+
+/// setup logger to be global between all threads
+pub struct Global;
+
+impl LogThreading for Global {
+    fn from_subscriber<S>(subscriber: S)
+    where
+        S: tracing::Subscriber + Send + Sync + 'static,
+    {
+        tracing::subscriber::set_global_default(subscriber)
+            .expect("setting default subscriber failed - have you already set a subscriber for the process globally?");
+    }
+}
+
+/// setup logger to be local to each thread.
+///
+/// Leaks the lock guard that is returned such that it lives for the duration of
+/// the program
+struct ThreadLocal;
+
+impl LogThreading for ThreadLocal {
+    fn from_subscriber<S>(subscriber: S)
+    where
+        S: tracing::Subscriber + Send + Sync + 'static,
+    {
+        let guard = tracing::subscriber::set_default(subscriber);
+        Box::leak(Box::new(guard));
+    }
+}
+
 /// setup logging with a specific [`LoggingOutput`] configuration
 ///
 /// ## Parameters
 ///
 /// `with_filename`: enable filename in logs
-pub fn logger_cfg(logging_output: LoggingOutput, with_filename: bool) {
+pub fn logger_cfg<LOG>(logging_output: LoggingOutput, with_filename: bool)
+where
+    LOG: LogThreading,
+{
     use tracing_subscriber::fmt::time;
 
     let time = time::SystemTime;
@@ -52,8 +102,8 @@ pub fn logger_cfg(logging_output: LoggingOutput, with_filename: bool) {
                 .with_timer(time)
                 .finish();
 
-            tracing::subscriber::set_global_default(subscriber)
-                .expect("setting default subscriber failed");
+            // initialize the logger to the given subscriber
+            LOG::from_subscriber(subscriber)
         }
         LoggingOutput::StdoutAndFile(file_writer) => {
             use tracing_subscriber::{fmt, prelude::*, registry::Registry};
@@ -74,8 +124,8 @@ pub fn logger_cfg(logging_output: LoggingOutput, with_filename: bool) {
                 .with(file_subscriber)
                 .with(stdout_subscriber);
 
-            tracing::subscriber::set_global_default(subscriber)
-                .expect("setting default subscriber failed");
+            // initialize the logger to the given subscriber
+            LOG::from_subscriber(subscriber)
         }
         LoggingOutput::File(file_writer) => {
             let subscriber = tracing_subscriber::FmtSubscriber::builder()
[…]
