### 2023-10-19 73faad3
fix clippy lints and compiler warnings
