+++ src/protocol/compiling.rs
@@ -56,6 +56,7 @@ impl Machine<Building, ClientBuildingState> {
     ///
     /// this routine is also responsible for listening for cancellation requests from the server
     #[instrument(skip(self), fields(batch_name=self.state.build_info.batch_name))]
+    #[allow(clippy::type_complexity)]
     pub(crate) async fn build_job(
         mut self,
     ) -> Result<
@@ -196,6 +197,7 @@ impl Machine<Building, ServerBuildingState> {
             batch_name = self.state.batch_name,
         )
     )]
+    #[allow(clippy::type_complexity)]
     pub(crate) async fn prepare_for_execution(
         mut self,
     ) -> Result<
