+++ src/protocol/compiling.rs
@@ -190,7 +190,7 @@ impl Machine<Building, ServerBuildingState> {
     ///
     /// this routine is also responsible for listening for cancellation requests from the server
     #[instrument(
-        skip(self), 
+        skip(self),
         fields(
             node_meta = %self.state.common.node_meta,
             namespace = self.state.namespace,
