### 2023-10-19 6b20636
fix Slurm constructor needed for python bindings
