+++ Cargo.toml
@@ -18,7 +18,7 @@ walkdir = { version = "2.3.2", optional = true }
 tokio = { version = "1.28.2", features = ["rt", "macros", "net", "io-util", "time", "sync", "fs", "process", "rt-multi-thread"], optional = true } 
 chrono = { version = "0.4.26", optional = true }
 futures = { version = "0.3.28", optional = true}
-nix = { version = "0.26.1", optional = true }
+nix = { version = "0.27.1", features = ["signal"], optional = true }
 num_cpus = { version = "1.15.0", optional = true }
 tokio-stream = { version = "0.1.14", features = ["fs"], optional = true}
 sha1 = { version = "0.10.5", optional = true }
@@ -37,7 +37,7 @@ git = "https://github.com/fluid-Dynamics-Group/matrix-notify"
 rev = "d88dd6518ec3ce1e5bb1992ecf1e229be5b98016"
 
 [dependencies.pyo3]
-version = "0.17.3" 
+version = "0.20.0" 
 features = ["extension-module"]
 optional = true
 
