### 2024-02-13 #5843 @VanillaBrooks
Hi all,

Sorry for delays on my response. A lot has changed in my life since I initially submitted this project for review and its difficult to find time to maintain it (see my github OSS contribution graph drop off a cliff in ~July).

For me personally, I no longer execute the scale of simulations that this project caters to. On top of that, the secondary consumers of this project, my research group, found their needs were better met by a battle tested workload managers (SLURM) on a traditional HPC cluster. I think this is telling with regards to future adoption of the tool.

I do think this project is interesting and has a niche, but I simply cannot support it in any reasonable time frame. Thank you @HaoZeke @jeremylt and @atrisovic for your time reviewing the submission! Given the above I think it best to close this issue for the foreseeable future - I would dislike publishing abandonware.
